requirements.txt
# sparrow
langchain   #~=0.1.16
langchain_experimental
langchain_core
langchain_cohere
langgraph
flask~=3.0.3
flask_socketio 
aiohttp~=3.9.5
requests~=2.31.0
websockets
bs4~=0.0.2
googlesearch-python
selenium~=4.20.0
yfinance~=0.2.38
psycopg2-binary
python-dotenv
asyncio

# generator
faker
langchain_ollama

# kgraph
semantic_text_splitter

# remember
mem0ai[graph]
sentence-transformers

# route
# google-maps-routeoptimization
# google-auth
google-api--python-client 
google-auth-oauthlib 
google-auth-httplib2
osmnx
matplotlib






# chromedriver
# setuptools~=68.2.0
# pytz~=2024.1
# beautifulsoup4~=4.12.3

# langchain-anthropic
# langchain-community
# duckduckgo-search
# langchain-ollama
# pydantic

# langchain-fireworks
# langchain-mistralai
# langchain-google-vertexai
# langchain-openai

# matplotlib~=3.8.4
# numpy~=1.26.4
# pandas~=2.2.2
# langchain-mongodb
# chromium-driver

# # starling
# python-docx
# transformers~=4.39.0
# elasticsearch>=7.0.0,<8.0.0
# requests~=2.31.0
# scikit-learn~=1.4.1.post1
# torch~=2.2.1
# apscheduler~=3.10.4
# feedparser~=6.0.11
# bs4~=0.0.2
# cryptography
# pypdf~=4.1.0

# # R2R
# r2r
# jwt
# supabase
# asyncpg
# aiofiles
# pdf2image
# unstructured_client
# psutil
# hatchet_sdk

